Added swagger api documentation
