+++ server/src/admin/routes/controls.js
@@ -3,35 +3,7 @@ import { banRecruiter, deleteJob, removeRecruiter, removeStudent, verifyRecruite
 
 
 const router = Router();
-/**
- * @swagger
- * tags:
- *   name: Admin
- *   description: Admin management API
-- */
 
-/**
- * @swagger
- * /verify-recruiter:
- *   post:
- *     summary: Verify a recruiter
- *     tags: [Admin]
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               recruiterId:
- *                 type: string
- *                 description: ID of the recruiter to verify
- *     responses:
- *       200:
- *         description: Recruiter verified successfully
- *       400:
- *         description: Bad request
- */
 router.post('/verify-recruiter',verifyRecruiter);
 router.post('/remove-recruiter',removeRecruiter);
 router.post('/remove-student',removeStudent);
